+++ examples/resolve_packages.rs
@@ -20,7 +20,7 @@ static FONT: &[u8] = include_bytes!("./fonts/texgyrecursor-regular.otf");
 
 #[cfg(feature = "packages")]
 fn main() {
-    let font = Font::new(Bytes::from(FONT), 0).expect("Could not parse font!");
+    let font = Font::new(Bytes::new(FONT.to_vec()), 0).expect("Could not parse font!");
 
     // Read in fonts and the main source file.
     // We can use this template more than once, if needed (Possibly